Move the session lifecycle tests into their own file
Message dispatch and filtering, the combat and zone teardown paths,
concurrent send plus receive, what happens after a session ends, and the
status and logging surfaces.

The Edge cases block lands here rather than with convergence, which is where
the split plan had put it sight unseen. Reading it settled the question: it
is wrong-sender rejection, guild filtering, decode warnings and DisableSync
teardown, which is this file's subject and not the other's.
